Add files via upload
Annotate 6 to 7 images (you can use attached cat_dog.zip for this and annotate cats and dogs using rectangles) using via annotator (find on internet or use the provided one in attachments), save the project, it will save all the annotations to a json file. Now your task is to create a notebook which load data from that json, load the information of bounding boxes along with label names and display / draw those on their respective images.